Move `DataSourceFactory.createFromConfig` out of api module into the only use site module
