+++ components/db/db-connection-manager-impl/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     implementation project(":components:db:db-connection-manager")
     implementation project(':libs:db:db-core')
     implementation project(':libs:configuration:configuration-core')
+    implementation project(':libs:configuration:configuration-validation')
     implementation project(':libs:configuration:configuration-datamodel')
     implementation project(':libs:virtual-node:virtual-node-datamodel')
     implementation project(":libs:lifecycle:lifecycle")
