+++ components/db/db-connection-manager-impl/src/main/kotlin/net/corda/db/connection/manager/impl/DbConnectionRepositoryFactory.kt
@@ -1,7 +1,7 @@
 package net.corda.db.connection.manager.impl
 
 import net.corda.db.connection.manager.DbConnectionsRepository
-import net.corda.db.connection.manager.createFromConfig
+import net.corda.db.connection.manager.impl.DataSourceFactoryHelper.createFromConfig
 import net.corda.db.core.CloseableDataSource
 import net.corda.db.core.DataSourceFactory
 import net.corda.libs.configuration.SmartConfig
